Fixed bug in merger, add log file for recording execution time + final file dimension

+++ src/main/java/it/unipi/dii/aide/mircv/models/CollectionStatistics.java
@@ -56,7 +56,7 @@ public class CollectionStatistics {
 
         channel.position(channel.size());
 
-        buffer.putLong(docCount);
+        buffer.putInt(docCount);
         buffer.putLong(totalLength);
 
         buffer = ByteBuffer.wrap(buffer.array());
